Extracted backgroundTaskToggle to its own component
